Change layout and add journey section
As the idea of the map for both experiences and education emerged,
I needed to remove one section, which is best as 6 was too long.
It was then necessary to redesign the contact section it was now an odd section.
Finally the map and some content for the journey section was added.
Map was made on boxy-svg.com
The idea is now to make appear only the information relevant to the hovered place on the map,
but it will probably be much more easy to do so in JS than in CSS,
which knowledge I lack at the moment, but it will come soon.